+++ dataflow/src/main/java/org/checkerframework/dataflow/util/PurityChecker.java
@@ -49,6 +49,7 @@ public class PurityChecker {
    * @param annoProvider the annotation provider
    * @param assumeSideEffectFree true if all methods should be assumed to be @SideEffectFree
    * @param assumeDeterministic true if all methods should be assumed to be @Deterministic
+   * @param assumePureGetters true if all getter methods should be assumed to be @Pure
    * @return information about whether the given statement is side-effect-free, deterministic, or
    *     both
    */
@@ -56,9 +57,11 @@ public class PurityChecker {
       TreePath statement,
       AnnotationProvider annoProvider,
       boolean assumeSideEffectFree,
-      boolean assumeDeterministic) {
+      boolean assumeDeterministic,
+      boolean assumePureGetters) {
     PurityCheckerHelper helper =
-        new PurityCheckerHelper(annoProvider, assumeSideEffectFree, assumeDeterministic);
+        new PurityCheckerHelper(
+            annoProvider, assumeSideEffectFree, assumeDeterministic, assumePureGetters);
     helper.scan(statement, null);
     return helper.purityResult;
   }
@@ -205,19 +208,28 @@ public class PurityChecker {
     private final boolean assumeDeterministic;
 
     /**
+     * True if all getter methods should be assumed to be @SideEffectFree and @Deterministic, for
+     * the purposes of org.checkerframework.dataflow analysis.
+     */
+    private final boolean assumePureGetters;
+
+    /**
      * Create a PurityCheckerHelper.
      *
      * @param annoProvider the annotation provider
      * @param assumeSideEffectFree true if all methods should be assumed to be @SideEffectFree
      * @param assumeDeterministic true if all methods should be assumed to be @Deterministic
+     * @param assumePureGetters true if getter methods should be assumed to be @Pure
      */
     public PurityCheckerHelper(
         AnnotationProvider annoProvider,
         boolean assumeSideEffectFree,
-        boolean assumeDeterministic) {
+        boolean assumeDeterministic,
+        boolean assumePureGetters) {
       this.annoProvider = annoProvider;
       this.assumeSideEffectFree = assumeSideEffectFree;
       this.assumeDeterministic = assumeDeterministic;
+      this.assumePureGetters = assumePureGetters;
     }
 
     @Override
@@ -237,7 +249,8 @@ public class PurityChecker {
         purityResult.addNotBothReason(tree, "call");
       } else {
         EnumSet<Pure.Kind> purityKinds =
-            (assumeDeterministic && assumeSideEffectFree)
+            ((assumeDeterministic && assumeSideEffectFree)
+                    || (assumePureGetters && ElementUtils.isGetter(elt)))
                 // Avoid computation if not necessary
                 ? detAndSeFree
                 : PurityUtils.getPurityKinds(annoProvider, elt);
@@ -294,6 +307,7 @@ public class PurityChecker {
       boolean deterministic =
           assumeDeterministic
               || okThrowDeterministic
+              // No need to check assumePureGetters because a constructor is never a getter.
               || PurityUtils.isDeterministic(annoProvider, ctorElement);
       boolean sideEffectFree =
           assumeSideEffectFree || PurityUtils.isSideEffectFree(annoProvider, ctorElement);