+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java
@@ -379,6 +379,12 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    */
   private final boolean assumeDeterministic;
 
+  /**
+   * True if all getter methods should be assumed to be @Pure, for the purposes of
+   * org.checkerframework.dataflow analysis.
+   */
+  private final boolean assumePureGetters;
+
   /** True if -AmergeStubsWithSource was provided on the command line. */
   private final boolean mergeStubsWithSource;
 
@@ -571,6 +577,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
         checker.hasOption("assumeSideEffectFree") || checker.hasOption("assumePure");
     this.assumeDeterministic =
         checker.hasOption("assumeDeterministic") || checker.hasOption("assumePure");
+    this.assumePureGetters = checker.hasOption("assumePureGetters");
 
     this.trees = Trees.instance(processingEnv);
     this.elements = processingEnv.getElementUtils();
@@ -5800,7 +5807,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
 
   @Override
   public boolean isSideEffectFree(ExecutableElement methodElement) {
-    if (assumeSideEffectFree) {
+    if (assumeSideEffectFree || (assumePureGetters && ElementUtils.isGetter(methodElement))) {
       return true;
     }
     if (ElementUtils.isRecordAccessor(methodElement)
@@ -5819,7 +5826,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
 
   @Override
   public boolean isDeterministic(ExecutableElement methodElement) {
-    if (assumeDeterministic) {
+    if (assumeDeterministic || (assumePureGetters && ElementUtils.isGetter(methodElement))) {
       return true;
     }
     if (ElementUtils.isRecordAccessor(methodElement)