+++ docs/CHANGELOG.md
@@ -10,9 +10,9 @@ New command-line options:
 
 Added method `isDeterministic()` to the `AnnotationProvider` interface.
 
-Marked `CFAbstractValue#leastUpperBound` and `CFAbstractValue#widenUpperBound`
-final.  Subclasses should override `CFAbstractValue#upperBound(V, TypeMirror,
-boolean)` instead.
+`CFAbstractValue#leastUpperBound` and `CFAbstractValue#widenUpperBound` are now
+final.  Subclasses should override method `CFAbstractValue#upperBound(V,
+TypeMirror, boolean)` instead.
 
 **Closed issues:**